Refactor fullset interpreter tests
